ES-1847: Db config creator
